# backend/openapi_baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PhantomNet Sentinel & Cyber Defense API",
    "description": "PhantomNet V3 -- AI-Driven Active Defense, Automated Incident Response Playbooks, MITRE ATT&CK Mapping, IDS Rule Generation, and TAXII 2.1 Threat Sharing Platform",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/events/{event_id}/pcap": {
      "get": {
        "tags": [
          "PCAP Analysis"
        ],
        "summary": "Download Pcap",
        "description": "Download the PCAP file associated with an event.",
        "operationId": "download_pcap_api_v1_events__event_id__pcap_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Event database ID",
              "title": "Event Id"
            },
            "description": "Event database ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pcap/analysis/{event_id}": {
      "get": {
        "tags": [
          "PCAP Analysis"
        ],
        "summary": "Get Pcap Analysis",
        "description": "Return the deep packet analysis for a given event's PCAP.",
        "operationId": "get_pcap_analysis_api_v1_pcap_analysis__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Event database ID",
              "title": "Event Id"
            },
            "description": "Event database ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/pcap/capture/{event_id}": {
      "post": {
        "tags": [
          "PCAP Analysis"
        ],
        "summary": "Trigger Capture",
        "description": "Manually trigger a packet capture for an event.",
        "operationId": "trigger_capture_api_v1_pcap_capture__event_id__post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Event database ID",
              "title": "Event Id"
            },
            "description": "Event database ID"
          },
          {
            "name": "duration",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3600,
              "minimum": 1,
              "description": "Capture duration in seconds",
              "default": 60,
              "title": "Duration"
            },
            "description": "Capture duration in seconds"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pcap/capture/{event_id}/status": {
      "get": {
        "tags": [
          "PCAP Analysis"
        ],
        "summary": "Capture Status",
        "description": "Check the status of an active or completed capture.",
        "operationId": "capture_status_api_v1_pcap_capture__event_id__status_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Event database ID",
              "title": "Event Id"
            },
            "description": "Event database ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pcap/cleanup": {
      "post": {
        "tags": [
          "PCAP Analysis"
        ],
        "summary": "Run Cleanup",
        "description": "Manually trigger PCAP retention cleanup.",
        "operationId": "run_cleanup_api_v1_pcap_cleanup_post",
        "parameters": [
          {
            "name": "retention_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "Retention window in days",
              "default": 30,
              "title": "Retention Days"
            },
            "description": "Retention window in days"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}